+++ containup/stack/stack.py
@@ -126,16 +126,40 @@ class Stack:
 
         self.mounts: dict[str, Volume] = {}
         self.networks: dict[str, Network] = {}
-        self.services: dict[str, Service] = {}
+        self.services: list[Service] = []
 
     def add(self, item_or_list: Union[StockItem, List[StockItem]]):
         items = item_or_list if isinstance(item_or_list, list) else [item_or_list]
         for item in items:
             logger.debug(item)
             if isinstance(item, Service):
-                self.services[item.name] = item
+                self.services.append(item)
             elif isinstance(item, Volume):
                 self.mounts[item.name] = item
             elif isinstance(item, Network):  # type: ignore
                 self.networks[item.name] = item
         return self
+
+    def filter_services(
+        self, filter_services: Optional[List[str]] = None
+    ) -> list[Service]:
+        """
+        Returns a list of services filtered by filter_services.
+
+        If filter_services is None, the services will not be filtered.
+        If filter_services is an empty list, the services will not be filtered.
+        Otherwise, the services returned should match the given filter_services list.
+
+        Arguments:
+            filter_services (list[str]) list of names of services
+        """
+        # if filter_services is empty or None then ignore it
+        # otherwise filter services to run to match filter_services (only the services the user wants to run)
+        targets: list[Service] = (
+            self.services
+            if not filter_services
+            else [
+                service for service in self.services if service.name in filter_services
+            ]
+        )
+        return targets